pyproject.toml
[tool.poetry]
name = "soci-daemon"
version = "0.1.0"
description = ""
authors = ["plebbimon <sebastiansmaland@gmail.com>"]
readme = "README.md"
packages = [{include = "soci_daemon"}]

[tool.poetry.dependencies]
python = "^3.10"
dbus-python = "^1.3.2"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
